rename dir and file
