+++ pom.xml
@@ -3,9 +3,9 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.yuyuyzl.SSS</groupId>
+    <groupId>com.sss.provider</groupId>
     <!-- TODO 修改这里的artifactId为合适的ID（不过也就只有自己看）-->
-    <artifactId>test-provider</artifactId>
+    <artifactId>sss-provider</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
 
@@ -198,7 +198,7 @@
         <!-- **************************** hibernate 依赖 **************************** -->
         <!-- **************************** interface 依赖 **************************** -->
         <dependency>
-            <groupId>com.yuyuyzl.SSS</groupId>
+            <groupId>com.sss.interfaces</groupId>
             <artifactId>sss-interfaces</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
