+++ src/main/resources/spring.xml
@@ -11,7 +11,7 @@
        http://www.springframework.org/schema/context
        http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <context:component-scan base-package="com.yuyuyzl.SSS"/>
+    <context:component-scan base-package="com.sss.provider"/>
 
     <import resource="spring-provider.xml"/>
